pyproject.toml
[project]
name = "ee-fin"
version = "0.1.0"
description = "Pre-experiment on explanatory equilibrium with bounded verification in a finance-themed setting."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "pydantic>=2.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]
openai = ["openai>=1.40.0"]
gemini = ["google-genai>=0.3.0"]

[tool.pytest.ini_options]
addopts = "-q"
